추천 친구 카드 수정
