+++ src/pages/Friend/Recommend.module.css
@@ -1,0 +1,31 @@
+.recommend {
+    display: grid;
+    grid-template-columns: repeat(auto-fit, minmax(220px, 1fr));
+    grid-row-gap: 20px;
+    justify-content: center;
+    gap: 30px;
+    justify-items: center;
+}
+
+.presentImg {
+    text-align: center;
+    width: 100%;
+    padding-top: 50px;
+}
+
+.vibration {
+    animation: vibration .1000ms infinite;
+}
+
+@keyframes vibration {
+    from {  
+        transform: rotate(1deg);
+    }
+    to {
+        transform: rotate(-1deg);
+    }
+}
+
+.img {
+    width: 200px;
+}
